requirements.txt
praw>=7.8.1
openai>=1.99.0
pyyaml>=6.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

